a little C library cleaning

+++ commandline/hidapi/mac/hid.c
@@ -327,7 +327,7 @@ static int get_string_property_utf8(IOHIDDeviceRef device, CFStringRef prop, cha
 			&used_buf_len);
 
 		buf[chars_copied] = 0;
-		return used_buf_len;
+		return (int)used_buf_len;
 	}
 	else
 		return 0;
@@ -822,7 +822,7 @@ static int set_report(hid_device *dev, IOHIDReportType type, const unsigned char
 					   data_to_send, length_to_send);
 	
 		if (res == kIOReturnSuccess) {
-			return length;
+			return (int)length;
 		}
 		else {
             // error could be 0xe000404f == kIOUSBPipeStalled
@@ -850,7 +850,7 @@ static int return_data(hid_device *dev, unsigned char *data, size_t length)
 	dev->input_reports = rpt->next;
 	free(rpt->data);
 	free(rpt);
-	return len;
+	return (int)len;
 }
 
 static int cond_wait(const hid_device *dev, pthread_cond_t *cond, pthread_mutex_t *mutex)
@@ -1000,7 +1000,7 @@ int HID_API_EXPORT hid_get_feature_report(hid_device *dev, unsigned char *data, 
 	                           data[0], /* Report ID */
 	                           data, &len);
 	if (res == kIOReturnSuccess)
-		return len;
+		return (int)len;
 	else
 		return -1;
 }
